bert step
cloud master: 1200.77 ms
cloud this: 1110.79 ms
native: 1114.81 ms
8% slower => within the margin of error

resnet step
cloud master: 13401.95 ms
cloud this: 3077.48 ms
native: 2131.50 ms
(530% slower => 44% slower, prob because a ton of stuff is being copied)

llama
cloud master: 10.44 tok/s
cloud this: 21.30 tok/s
native: 23.01 tok/s
(120% slower => 8% slower)

It also fixes the [weird ci thing](https://discord.com/channels/1068976834382925865/1294150613738983445/1370223404355162225) @chenyuxyz found.

master:
<img width="1333" alt="image" src="https://github.com/user-attachments/assets/47ed16fb-4688-484d-9eaf-9ccce43a72f6" />
this:
<img width="1342" alt="image" src="https://github.com/user-attachments/assets/5d125982-f137-4894-99cb-748e7d258952" />

4 minutes => 25 seconds total time


And it's capable of handling more than one connection (python http server just doesn't accept new connections before the old one is closed) which will allow multiple clients to connect at the same time and infiniband discovery server to not be a separate thing